WIP refactor approve and reject tx modal
