+++ src/routes/safe/components/Transactions/TxList/modals/RejectTxModal.tsx
@@ -2,8 +2,7 @@ import { MultisigExecutionInfo } from '@gnosis.pm/safe-react-gateway-sdk'
 
 import { useDispatch } from 'react-redux'
 import { useStyles } from './style'
-import Modal, { ButtonStatus, Modal as GenericModal } from 'src/components/Modal'
-import { ReviewInfoText } from 'src/components/ReviewInfoText'
+import Modal from 'src/components/Modal'
 import Block from 'src/components/layout/Block'
 import Bold from 'src/components/layout/Bold'
 import Hairline from 'src/components/layout/Hairline'
@@ -13,14 +12,11 @@ import { TX_NOTIFICATION_TYPES } from 'src/logic/safe/transactions'
 import { EMPTY_DATA } from 'src/logic/wallets/ethTransactions'
 import { createTransaction } from 'src/logic/safe/store/actions/createTransaction'
 import { Transaction } from 'src/logic/safe/store/models/types/gateway.d'
-import { EstimationStatus, useEstimateTransactionGas } from 'src/logic/hooks/useEstimateTransactionGas'
-import { TxParametersDetail } from 'src/routes/safe/components/Transactions/helpers/TxParametersDetail'
-import { EditableTxParameters } from 'src/routes/safe/components/Transactions/helpers/EditableTxParameters'
 import { TxParameters } from 'src/routes/safe/container/hooks/useTransactionParameters'
 import { ParametersStatus } from 'src/routes/safe/components/Transactions/helpers/utils'
 import { ModalHeader } from 'src/routes/safe/components/Balances/SendModal/screens/ModalHeader'
 import { extractSafeAddress } from 'src/routes/routes'
-import useCanTxExecute from 'src/logic/hooks/useCanTxExecute'
+import { TxModalWrapper } from 'src/routes/safe/components/Transactions/helpers/TxModalWrapper'
 
 type Props = {
   isOpen: boolean
@@ -32,19 +28,6 @@ export const RejectTxModal = ({ isOpen, onClose, gwTransaction }: Props): React.
   const dispatch = useDispatch()
   const safeAddress = extractSafeAddress()
   const classes = useStyles()
-
-  const {
-    gasCostFormatted,
-    txEstimationExecutionStatus,
-    isOffChainSignature,
-    isCreation,
-    gasLimit,
-    gasPriceFormatted,
-  } = useEstimateTransactionGas({
-    txData: EMPTY_DATA,
-    txRecipient: safeAddress,
-  })
-  const canTxExecute = useCanTxExecute()
 
   const origin = gwTransaction.safeAppInfo
     ? JSON.stringify({ name: gwTransaction.safeAppInfo.name, url: gwTransaction.safeAppInfo.url })
@@ -71,78 +54,31 @@ export const RejectTxModal = ({ isOpen, onClose, gwTransaction }: Props): React.
   const getParametersStatus = (): ParametersStatus => {
     return 'CANCEL_TRANSACTION'
   }
-
-  let confirmButtonStatus: ButtonStatus = ButtonStatus.READY
-  let confirmButtonText = 'Reject transaction'
-  if (txEstimationExecutionStatus === EstimationStatus.LOADING) {
-    confirmButtonStatus = ButtonStatus.LOADING
-    confirmButtonText = 'Estimating'
-  }
-
   return (
     <Modal description="Reject transaction" handleClose={onClose} open={isOpen} title="Reject Transaction">
-      <EditableTxParameters
-        isOffChainSignature={isOffChainSignature}
-        isExecution={canTxExecute}
-        ethGasLimit={gasLimit}
-        ethGasPrice={gasPriceFormatted}
-        safeTxGas={'0'}
-        safeNonce={nonce.toString()}
+      <TxModalWrapper
+        txData={EMPTY_DATA}
+        onSubmit={sendReplacementTransaction}
+        onBack={onClose}
+        submitText="Reject Transaction"
         parametersStatus={getParametersStatus()}
       >
-        {(txParameters, toggleEditMode) => {
-          return (
-            <>
-              <ModalHeader onClose={onClose} title="Reject transaction" />
-              <Hairline />
-              <Block className={classes.container}>
-                <Row>
-                  <Paragraph>
-                    This action will reject this transaction. A separate transaction will be performed to submit the
-                    rejection.
-                  </Paragraph>
-                  <Paragraph color="medium" size="sm">
-                    Transaction nonce:
-                    <br />
-                    <Bold className={classes.nonceNumber}>{nonce}</Bold>
-                  </Paragraph>
-                </Row>
-                {/* Tx Parameters */}
-                <TxParametersDetail
-                  txParameters={txParameters}
-                  onEdit={toggleEditMode}
-                  parametersStatus={getParametersStatus()}
-                  isTransactionCreation={isCreation}
-                  isTransactionExecution={canTxExecute}
-                  isOffChainSignature={isOffChainSignature}
-                />
-              </Block>
-
-              {txEstimationExecutionStatus === EstimationStatus.LOADING ? null : (
-                <ReviewInfoText
-                  gasCostFormatted={gasCostFormatted}
-                  isCreation={isCreation}
-                  isExecution={canTxExecute}
-                  safeNonce={txParameters.safeNonce}
-                  txEstimationExecutionStatus={txEstimationExecutionStatus}
-                />
-              )}
-              <GenericModal.Footer withoutBorder={confirmButtonStatus !== ButtonStatus.LOADING}>
-                <GenericModal.Footer.Buttons
-                  cancelButtonProps={{ onClick: onClose, text: 'Close' }}
-                  confirmButtonProps={{
-                    onClick: () => sendReplacementTransaction(txParameters),
-                    color: 'error',
-                    type: 'submit',
-                    status: confirmButtonStatus,
-                    text: confirmButtonText,
-                  }}
-                />
-              </GenericModal.Footer>
-            </>
-          )
-        }}
-      </EditableTxParameters>
+        <ModalHeader onClose={onClose} title="Reject transaction" />
+        <Hairline />
+        <Block className={classes.container}>
+          <Row>
+            <Paragraph>
+              This action will reject this transaction. A separate transaction will be performed to submit the
+              rejection.
+            </Paragraph>
+            <Paragraph color="medium" size="sm">
+              Transaction nonce:
+              <br />
+              <Bold className={classes.nonceNumber}>{nonce}</Bold>
+            </Paragraph>
+          </Row>
+        </Block>
+      </TxModalWrapper>
     </Modal>
   )
 }
